+++ packages/expo/src/dom/dom.ts
@@ -1,6 +1,7 @@
 // Native file
 
-export type { DOMProps } from './dom.types';
+export * from './dom-hooks';
+export type { DOMProps, DOMImperativeFactory } from './dom.types';
 
 // TODO: Maybe this could be a bundler global instead.
 /** @returns `true` when the current JS running in a DOM Component environment. */